[Fix] Fixed incorrect datatype for neural net learning rate

+++ lib/headers/neural_net.h
@@ -55,7 +55,7 @@ typedef struct NeuralNetOpt {
     // The initial value for the biases of the layers.
     double initialBias;
     // The learning rate of the neural network.
-    int lr;
+    double lr;
 } NeuralNetOpt;
 
 /** @brief Stucture for the NeuralNetwork. */
